Redesigned plugin api-response handling overview
This may seem like a step backwards, but it's necessary to give plugin
devs more control over what to do in response to an api call.

Additionally, this is necessary for the wallet changes that now use
POST, GET, and PUT type calls on the same endpoint for different
functionalities.