don't update if a file is being renamed
kind of an ugly hack; nothing at all will be updated until the user
finishes renaming.
also, there are some usability issues if the window loses focus. The
name will no longer appear to be editable, but it will be, and this
prevents updating until the user clicks somewhere in the plugin.